requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
Pillow>=10.0,<11.0
PyPDF2>=3.0,<4.0
google-generativeai>=0.5,<1.0